refactor(settings): split import-panel into per-source card components
