+++ src/components/settings/import-panel/json-import-card.tsx
@@ -1,0 +1,235 @@
+"use client";
+
+import { useId, useRef, useState } from "react";
+import Link from "next/link";
+import {
+  AlertCircle,
+  CheckCircle2,
+  Download,
+  FileJson,
+  Loader2,
+  Upload,
+} from "lucide-react";
+
+import { Button } from "@/components/ui/button";
+import { Label } from "@/components/ui/label";
+import { Textarea } from "@/components/ui/textarea";
+import { useTranslations } from "@/lib/i18n/context";
+import { apiFetchRaw } from "@/lib/api/api-fetch";
+import { ImportCardShell } from "./import-card-shell";
+import { MAX_PASTE_CHARS } from "./constants";
+import { EXAMPLE_IMPORT, parseImportJson } from "./import-examples";
+
+interface JsonImportResult {
+  measurements: number;
+  moodEntries: number;
+  skipped: number;
+}
+
+export function JsonImportCard() {
+  const { t } = useTranslations();
+  const fileInputRef = useRef<HTMLInputElement>(null);
+  const [text, setText] = useState("");
+  const [busy, setBusy] = useState(false);
+  const [error, setError] = useState<string | null>(null);
+  const [result, setResult] = useState<JsonImportResult | null>(null);
+  const textareaId = useId();
+
+  async function onFileChange(e: React.ChangeEvent<HTMLInputElement>) {
+    const file = e.target.files?.[0];
+    e.target.value = "";
+    if (!file) return;
+    setError(null);
+    setResult(null);
+    try {
+      const content = await file.text();
+      setText(content);
+    } catch {
+      setError(t("settings.sections.export.import.json.readFailed"));
+    }
+  }
+
+  function downloadExample() {
+    const blob = new Blob([JSON.stringify(EXAMPLE_IMPORT, null, 2)], {
+      type: "application/json",
+    });
+    const url = URL.createObjectURL(blob);
+    const a = document.createElement("a");
+    a.href = url;
+    a.download = "healthlog-import-example.json";
+    document.body.appendChild(a);
+    a.click();
+    a.remove();
+    URL.revokeObjectURL(url);
+  }
+
+  async function handleImport() {
+    setError(null);
+    setResult(null);
+    const parsed = parseImportJson(text);
+    if (!parsed.ok) {
+      setError(t("settings.sections.export.import.json.invalidJson"));
+      return;
+    }
+    const payload = parsed.value;
+    setBusy(true);
+    try {
+      const res = await apiFetchRaw("/api/import", {
+        method: "POST",
+        headers: { "Content-Type": "application/json" },
+        credentials: "include",
+        body: JSON.stringify(payload),
+      });
+      if (res.status === 429) {
+        setError(t("settings.sections.export.import.json.rateLimited"));
+        return;
+      }
+      if (res.status === 422) {
+        const body = await res.json().catch(() => null);
+        setError(
+          body?.error ??
+            t("settings.sections.export.import.json.invalidPayload"),
+        );
+        return;
+      }
+      if (!res.ok) {
+        setError(t("settings.sections.export.import.json.failed"));
+        return;
+      }
+      const data = (await res.json()).data as JsonImportResult;
+      setResult({
+        measurements: data?.measurements ?? 0,
+        moodEntries: data?.moodEntries ?? 0,
+        skipped: data?.skipped ?? 0,
+      });
+    } catch {
+      setError(t("settings.sections.export.import.json.failed"));
+    } finally {
+      setBusy(false);
+    }
+  }
+
+  return (
+    <ImportCardShell
+      testId="import-card-json"
+      icon={FileJson}
+      title={t("settings.sections.export.import.json.title")}
+      description={t("settings.sections.export.import.json.description")}
+    >
+      <div className="space-y-1.5">
+        <Label htmlFor={textareaId} className="text-xs">
+          {t("settings.sections.export.import.json.pasteLabel")}
+        </Label>
+        <Textarea
+          id={textareaId}
+          data-testid="import-json-textarea"
+          value={text}
+          onChange={(e) => setText(e.target.value)}
+          rows={5}
+          maxLength={MAX_PASTE_CHARS}
+          spellCheck={false}
+          placeholder='{"measurements":[…],"moodEntries":[…]}'
+          className="font-mono text-xs"
+        />
+        <p className="text-muted-foreground text-right text-xs tabular-nums">
+          {t("settings.sections.export.import.charCount", {
+            used: text.length,
+            max: MAX_PASTE_CHARS,
+          })}
+        </p>
+      </div>
+
+      <input
+        ref={fileInputRef}
+        type="file"
+        accept=".json,application/json"
+        className="sr-only"
+        aria-label={t("settings.sections.export.import.json.fileInputLabel")}
+        onChange={onFileChange}
+      />
+
+      <p className="text-muted-foreground text-xs">
+        {t("settings.sections.export.import.json.schemaHint")}{" "}
+        <Link
+          href="/docs/integrations/data-import"
+          className="text-primary underline underline-offset-2"
+        >
+          {t("settings.sections.export.import.json.docsLink")}
+        </Link>
+      </p>
+
+      <div aria-live="polite" className="space-y-2">
+        {result && (
+          <p
+            data-testid="import-json-result"
+            className="text-foreground flex items-start gap-2 text-xs"
+          >
+            <CheckCircle2
+              className="text-success mt-0.5 h-3.5 w-3.5 shrink-0"
+              aria-hidden="true"
+            />
+            <span>
+              {t("settings.sections.export.import.json.resultSummary", {
+                measurements: result.measurements,
+                moods: result.moodEntries,
+                skipped: result.skipped,
+              })}
+            </span>
+          </p>
+        )}
+        {error && (
+          <p
+            role="alert"
+            className="text-destructive flex items-start gap-2 text-sm"
+          >
+            <AlertCircle
+              className="mt-0.5 h-3.5 w-3.5 shrink-0"
+              aria-hidden="true"
+            />
+            <span>{error}</span>
+          </p>
+        )}
+      </div>
+
+      <div className="mt-auto flex flex-wrap items-center gap-3 pt-1">
+        <Button
+          type="button"
+          variant="outline"
+          size="sm"
+          className="min-h-11 sm:min-h-9"
+          onClick={() => fileInputRef.current?.click()}
+          data-testid="import-json-choose-file"
+        >
+          <Upload className="h-3.5 w-3.5" />
+          {t("settings.sections.export.import.json.uploadFile")}
+        </Button>
+        <Button
+          type="button"
+          variant="ghost"
+          size="sm"
+          className="min-h-11 sm:min-h-9"
+          onClick={downloadExample}
+          data-testid="import-json-download-example"
+        >
+          <Download className="h-3.5 w-3.5" />
+          {t("settings.sections.export.import.json.downloadExample")}
+        </Button>
+        <Button
+          type="button"
+          size="sm"
+          className="min-h-11 sm:min-h-9"
+          disabled={busy || text.trim().length === 0}
+          onClick={handleImport}
+          data-testid="import-action-json"
+        >
+          {busy ? (
+            <Loader2 className="h-3.5 w-3.5 animate-spin motion-reduce:animate-none" />
+          ) : (
+            <FileJson className="h-3.5 w-3.5" />
+          )}
+          {t("settings.sections.export.import.json.import")}
+        </Button>
+      </div>
+    </ImportCardShell>
+  );
+}
